[PackageBuilder] move tests as well
