--- packages/PackageBuilder/tests/Adapter/Source/CollectedInterface.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\PackageBuilder\Tests\Adapter\Source;
-
-interface CollectedInterface
-{
-}
