+++ packages/PackageBuilder/tests/ContainerFactory.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\PackageBuilder\Tests\Adapter\Symfony;
+namespace Symplify\PackageBuilder\Tests;
 
 use Psr\Container\ContainerInterface;
