+++ packages/PackageBuilder/tests/DependencyInjectionSource/CollectedInterface.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\PackageBuilder\Tests\DependencyInjectionSource;
+
+interface CollectedInterface
+{
+}
