+++ packages/PackageBuilder/tests/DependencyInjectionSource/CollectorInterface.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\PackageBuilder\Tests\Adapter\Source;
+namespace Symplify\PackageBuilder\Tests\DependencyInjectionSource;
 
 interface CollectorInterface
 {
